main.go: enable cleanup of old PVCs
This commit enables the controller to clean up the PVCs of Thanos
Receive Pods that are watched by it. Whenever a receiver is deleted, the
controller will spawn a helper container that mounts all PVCs specified
by the StatefulSet for that container and `rm -rf`s the contents of them.

Tested on a kind cluster. A follow up PR will add E2E tests and verify
this functionality.