Allow to annotate pods on change using a different label
This change allows users to optionally specify a different label (than the
one used to watch receiver pods) for annotating pods on hashring change.

This is useful in a separate Thanos receiver router and ingestor setup:
https://thanos.io/tip/proposals-accepted/202012-receive-split.md/ where
we need to watch a different set of ingestor pods to update the hashring,
but the hashring is consumed by a different set of router pods.

Signed-off-by: Ratnadeep Debnath <rtnpro@gmail.com>